requirements.txt
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.0.3
Flask-Cors==5.0.0
greenlet==3.1.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==2.1.5
praw==7.8.1
prawcore==2.4.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
SQLAlchemy==2.0.39
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.6
zipp==3.20.2
